Pushing AssignMessage common elements into config/policy/assign_message_elements.go
